# documentation/swagger.json-OpenApi3Json (1).json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Homework 9",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "http://localhost:3000/login",
        "operationId": "http://localhost:3000/login",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/httplocalhost3000loginrequest"
                  },
                  {
                    "example": {
                      "email": "oainger0@craigslist.org",
                      "password": "KcAk6Mrg7DRM"
                    }
                  }
                ]
              },
              "example": {
                "email": "oainger0@craigslist.org",
                "password": "KcAk6Mrg7DRM"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "register",
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/registerrequest"
                  },
                  {
                    "example": {
                      "email": "brahmantiyonico@gmail.com",
                      "password": "Shineclassic99",
                      "gender": "Male",
                      "role": "Software Engineer"
                    }
                  }
                ]
              },
              "example": {
                "email": "brahmantiyonico@gmail.com",
                "password": "Shineclassic99",
                "gender": "Male",
                "role": "Software Engineer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/": {
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "http://localhost:3000",
        "operationId": "http://localhost:3000",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/101": {
      "delete": {
        "tags": [
          "movies"
        ],
        "summary": "http://localhost:3000/3",
        "operationId": "http://localhost:3000/3",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/3": {
      "put": {
        "tags": [
          "movies"
        ],
        "summary": "http://localhost:3000/3",
        "operationId": "Puthttp://localhost:3000/3",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/httplocalhost30003request"
                  },
                  {
                    "example": {
                      "id": 3,
                      "genres": "Romance"
                    }
                  }
                ]
              },
              "example": {
                "id": 3,
                "genres": "Romance"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
      "httplocalhost3000loginrequest": {
        "title": "httplocalhost3000loginrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "oainger0@craigslist.org",
          "password": "KcAk6Mrg7DRM"
        }
      },
      "registerrequest": {
        "title": "registerrequest",
        "required": [
          "email",
          "password",
          "gender",
          "role"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "example": {
          "email": "brahmantiyonico@gmail.com",
          "password": "Shineclassic99",
          "gender": "Male",
          "role": "Software Engineer"
        }
      },
      "httplocalhost30003request": {
        "title": "httplocalhost30003request",
        "required": [
          "id",
          "genres"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "genres": {
            "type": "string"
          }
        },
        "example": {
          "id": 3,
          "genres": "Romance"
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
  "tags": [
    {
      "name": "Users"
    },
    {
      "name": "movies"
    }
  ]
}